backend/pyproject.toml
[project]
name = "momentum"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2,<6.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "pytz (>=2025.2,<2026.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "django-filter (>=25.1,<26.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "freezegun (>=1.5.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.7"

[tool.pytest.ini_options]
testpaths = ["momentum/*"]
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "-vv -p no:cacheprovider --ds=momentum.settings.test"
pythonpath = ". momentum"
filterwarnings = ["ignore::DeprecationWarning"]
